Terminado os Exercicios

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 // gerenciadores de rota
 const blogsRouter = require('express').Router();
+const logger = require('../utils/logger');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 const jwt = require('jsonwebtoken');
+const { userExtractor } = require('../utils/middleware');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
@@ -19,21 +21,16 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 });
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization');
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '');
-  }
-  return null;
-};
-
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body;
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
-  if (!decodedToken.id) {
+  const token = request.token;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!(token && decodedToken.id)) {
     return response.status(401).json({ error: 'token invalid' });
   }
-  const user = await User.findById(decodedToken.id);
+  // const user = await User.findById(decodedToken.id);
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -50,9 +47,17 @@ blogsRouter.post('/', async (request, response) => {
   response.json(savedBlog);
 });
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user;
+  const userId = user.id;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'Unauthorized' });
+  }
 });
 
 blogsRouter.put('/:id', (request, response, next) => {
